Added rotations for the player its shadow.
The player its shadow is a seperate object, because the player its global object (viewed by other players or possibly through mirrors) can't cast shadows when not rendered. Since the global object must be invisible in the local camera, there is now a seperate object which is visible both locally and globally, which fixes the issue.